pyproject.toml
[tool.poetry]
name = "logger_example"
version = "0.1.0"
description = ""
authors = ["MinDong <mdskylover@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
